cleanup and proxy highlight api methods
