+++ sdk/highlight-run/src/client/index.tsx
@@ -457,6 +457,17 @@ export class Highlight {
 			user_identifier.toString(),
 		)
 		setItem(SESSION_STORAGE_KEYS.USER_OBJECT, JSON.stringify(user_object))
+		this._ldClient?.identify({
+			kind: 'multi',
+			user: {
+				key: user_identifier,
+				identifier: user_identifier,
+			},
+			session: {
+				key: this.sessionData.sessionSecureID,
+				sessionSecureID: this.sessionData.sessionSecureID,
+			},
+		})
 		this._worker.postMessage({
 			message: {
 				type: MessageType.Identify,
@@ -532,6 +543,11 @@ export class Highlight {
 			} catch {
 				delete obj[key]
 			}
+		})
+		this._ldClient?.track('$ld:track', {
+			sessionSecureID: this.sessionData.sessionSecureID,
+			propertyType: typeArg,
+			...properties_obj,
 		})
 		this._worker.postMessage({
 			message: {
@@ -1512,7 +1528,7 @@ SessionSecureID: ${this.sessionData.sessionSecureID}`,
 		this._lastSnapshotTime = new Date().getTime()
 	}
 
-	register(client: LDClientMin) {
+	registerLD(client: LDClientMin) {
 		this._ldClient = client
 	}
 }
